+++ .src.legacy/webapp/components/Navigation/NavAssessment/Header/Title/Title.scss
@@ -1,0 +1,9 @@
+.nav-assessment-header__btn-lock {
+  display: flex;
+  border: none;
+
+  &:disabled {
+    border: none;
+    opacity: 1;
+  }
+}
